requirements.txt
streamlit==1.29.0
qrcode[pil]==7.3.1
pillow==9.4.0